# SIMD: Replace raw SIMD of sin/cos with NPYV(universal intrinsics)