Merge after #17790, #17789

#### SIMD: Replace raw SIMD of sin/cos with NPYV

   The new code improves the performance of non-contiguous memory access
   for the output array without any reduction in performance.
   For PPC64LE the performance increased by 2-3.0, and 1.5-2.0 on aarch64.

TODO:
- [x] remove temporary commits
- [x] benchmarks
- [x] add testing cases for the new intrinsics
- [x] rebase after #17790, #17789
- [x] re-run the performance tests and update the current one

# Performance tests(ASV)
#### Args
```
--bench-compare master bench_ufunc_strides.Unary -- --sort name --cpu-affinity 1,5
```

### X86
I had to count on my local machine because I couldn't able to get stable ratios using aws.
see standalone benchamrk for AVX512F.

<details>
<summary>CPU</summary>

```Bash

Architecture:        x86_64
CPU op-mode(s):      32-bit, 64-bit
Byte Order:          Little Endian
CPU(s):              8
On-line CPU(s) list: 0-7
Thread(s) per core:  2
Core(s) per socket:  4
Socket(s):           1
NUMA node(s):        1
Vendor ID:           GenuineIntel
CPU family:          6
Model:               142
Model name:          Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
Stepping:            10
CPU MHz:             1800.344
CPU max MHz:         4000.0000
CPU min MHz:         400.0000
BogoMIPS:            3984.00
Virtualization:      VT-x
L1d cache:           32K
L1i cache:           32K
L2 cache:            256K
L3 cache:            8192K
NUMA node0 CPU(s):   0-7
Flags:               fpu vme de pse tsc msr pae mce cx8 apic sep mtrr pge mca cmov pat pse36 clflush dts acpi mmx fxsr sse sse2 ss ht tm pbe syscall nx pdpe1gb rdtscp lm constant_tsc art arch_perfmon pebs bts rep_good nopl xtopology nonstop_tsc cpuid aperfmperf tsc_known_freq pni pclmulqdq dtes64 monitor ds_cpl vmx est tm2 ssse3 sdbg fma cx16 xtpr pdcm pcid sse4_1 sse4_2 x2apic movbe popcnt tsc_deadline_timer aes xsave avx f16c rdrand lahf_lm abm 3dnowprefetch cpuid_fault epb invpcid_single pti ssbd ibrs ibpb stibp tpr_shadow vnmi flexpriority ept vpid ept_ad fsgsbase tsc_adjust bmi1 avx2 smep bmi2 erms invpcid mpx rdseed adx
```
</details>

<details>
<summary>OS</summary>

```
Linux ac6279ab1a82 4.19.0-13-amd64 #1 SMP Debian 4.19.160-2 (2020-11-28) x86_64 x86_64 x86_64 GNU/Linux
gcc (Ubuntu 8.4.0-1ubuntu1~18.04) 8.4.0
```
</details>

#### Benchmark
<details>
 <summary>AVX2 & FMA3 - Changed only</summary>

```Bash
       before           after         ratio
     [098a3b41]       [a0322ee9]
     <master>         <to_npyv_sincos_f32>
        259~3us       55.1~0.2us     0.21  bench_ufunc_strides.Unary.time_ufunc('cos', 1, 2, 'f')
        260~4us       56.2~0.2us     0.22  bench_ufunc_strides.Unary.time_ufunc('cos', 1, 4, 'f')
      334~0.8us      60.4~0.07us     0.18  bench_ufunc_strides.Unary.time_ufunc('cos', 2, 2, 'f')
      335~0.9us       61.5~0.2us     0.18  bench_ufunc_strides.Unary.time_ufunc('cos', 2, 4, 'f')
      337~0.4us       62.1~0.2us     0.18  bench_ufunc_strides.Unary.time_ufunc('cos', 4, 2, 'f')
        339~2us       61.2~0.6us     0.18  bench_ufunc_strides.Unary.time_ufunc('cos', 4, 4, 'f')
       266~10us       54.9~0.2us     0.21  bench_ufunc_strides.Unary.time_ufunc('sin', 1, 2, 'f')
       270~20us       55.6~0.2us     0.21  bench_ufunc_strides.Unary.time_ufunc('sin', 1, 4, 'f')
        331~3us       60.3~0.1us     0.18  bench_ufunc_strides.Unary.time_ufunc('sin', 2, 2, 'f')
        332~2us       61.0~0.3us     0.18  bench_ufunc_strides.Unary.time_ufunc('sin', 2, 4, 'f')
        336~1us       61.7~0.3us     0.18  bench_ufunc_strides.Unary.time_ufunc('sin', 4, 2, 'f')
      335~0.2us       61.5~0.4us     0.18  bench_ufunc_strides.Unary.time_ufunc('sin', 4, 4, 'f')
```
</details>

### Power little-endian
<details>
<summary>CPU</summary>

```
Architecture:                    ppc64le
Byte Order:                      Little Endian
CPU(s):                          8
On-line CPU(s) list:             0-7
Thread(s) per core:              1
Core(s) per socket:              1
Socket(s):                       8
NUMA node(s):                    1
Model:                           2.2 (pvr 004e 1202)
Model name:                      POWER9 (architected), altivec supported
L1d cache:                       256 KiB
L1i cache:                       256 KiB
NUMA node0 CPU(s):               0-7
Vulnerability L1tf:              Not affected
Vulnerability Meltdown:          Mitigation; RFI Flush
Vulnerability Spec store bypass: Mitigation; Kernel entry/exit barrier (eieio)
Vulnerability Spectre v1:        Mitigation; __user pointer sanitization
Vulnerability Spectre v2:        Vulnerable

processor   : 7
cpu     : POWER9 (architected), altivec supported
clock       : 2200.000000MHz
revision    : 2.2 (pvr 004e 1202)

timebase    : 512000000
platform    : pSeries
model       : IBM pSeries (emulated by qemu)
machine     : CHRP IBM pSeries (emulated by qemu)
MMU     : Radix


```

</details>

<details>
<summary>OS</summary>

```
Linux 8b2db3b0dfac 4.19.0-2-powerpc64le
gcc version 9.2.1 20191008 (Ubuntu 9.2.1-9ubuntu2) 
```
</details>

#### Benchmark


<details>
 <summary>VSX2(ISA >= 2.07) - Changed only</summary>

```
       before           after         ratio
     [098a3b41]       [a0322ee9]
     <master>         <to_npyv_sincos_f32>
       120±0.2μs      44.7±0.02μs     0.37  bench_ufunc_strides.Unary.time_ufunc('cos', 1, 1, 'f')
       121±0.5μs      48.9±0.04μs     0.40  bench_ufunc_strides.Unary.time_ufunc('cos', 1, 2, 'f')
       121±0.3μs      49.1±0.04μs     0.41  bench_ufunc_strides.Unary.time_ufunc('cos', 1, 4, 'f')
       121±0.2μs      48.7±0.02μs     0.40  bench_ufunc_strides.Unary.time_ufunc('cos', 2, 1, 'f')
       121±0.1μs      52.4±0.04μs     0.43  bench_ufunc_strides.Unary.time_ufunc('cos', 2, 2, 'f')
       121±0.1μs      52.5±0.05μs     0.43  bench_ufunc_strides.Unary.time_ufunc('cos', 2, 4, 'f')
       121±0.2μs      48.8±0.06μs     0.40  bench_ufunc_strides.Unary.time_ufunc('cos', 4, 1, 'f')
       122±0.6μs      52.6±0.04μs     0.43  bench_ufunc_strides.Unary.time_ufunc('cos', 4, 2, 'f')
      122±0.09μs      53.0±0.01μs     0.43  bench_ufunc_strides.Unary.time_ufunc('cos', 4, 4, 'f')
       126±0.6μs      44.1±0.01μs     0.35  bench_ufunc_strides.Unary.time_ufunc('sin', 1, 1, 'f')
       131±0.5μs      48.2±0.02μs     0.37  bench_ufunc_strides.Unary.time_ufunc('sin', 1, 2, 'f')
       130±0.7μs      48.4±0.02μs     0.37  bench_ufunc_strides.Unary.time_ufunc('sin', 1, 4, 'f')
       131±0.6μs      47.9±0.02μs     0.37  bench_ufunc_strides.Unary.time_ufunc('sin', 2, 1, 'f')
       131±0.5μs      51.4±0.04μs     0.39  bench_ufunc_strides.Unary.time_ufunc('sin', 2, 2, 'f')
       131±0.6μs      51.6±0.02μs     0.39  bench_ufunc_strides.Unary.time_ufunc('sin', 2, 4, 'f')
       130±0.7μs      48.1±0.02μs     0.37  bench_ufunc_strides.Unary.time_ufunc('sin', 4, 1, 'f')
       131±0.2μs      51.7±0.02μs     0.39  bench_ufunc_strides.Unary.time_ufunc('sin', 4, 2, 'f')
       131±0.4μs      52.0±0.05μs     0.40  bench_ufunc_strides.Unary.time_ufunc('sin', 4, 4, 'f')
```

</details>


# Performance tests(standalone #15987)
#### Args used within #15987
```
--filter "(sin|cos)::.*[f]" --strides 1 2 3 10 --msleep 1 --iteration 100
```

Note: `--msleep 1` force the running thread to sleep 1 millisecond before collecting each sample
to revert any frequency reduction, since it seems that throttling effect on wall time when ``AVX512F`` is enabled.

### X86
<details>
<summary>CPU</summary>

```
Architecture:                    x86_64
CPU op-mode(s):                  32-bit, 64-bit
Byte Order:                      Little Endian
Address sizes:                   46 bits physical, 48 bits virtual
CPU(s):                          4
On-line CPU(s) list:             0-3
Thread(s) per core:              2
Core(s) per socket:              2
Socket(s):                       1
NUMA node(s):                    1
Vendor ID:                       GenuineIntel
CPU family:                      6
Model:                           85
Model name:                      Intel(R) Xeon(R) Platinum 8275CL CPU @ 3.00GH
                                 z
Stepping:                        7
CPU MHz:                         3604.410
BogoMIPS:                        6000.00
Hypervisor vendor:               KVM
Virtualization type:             full
L1d cache:                       64 KiB
L1i cache:                       64 KiB
L2 cache:                        2 MiB
L3 cache:                        35.8 MiB
NUMA node0 CPU(s):               0-3
Vulnerability Itlb multihit:     KVM: Vulnerable
Vulnerability L1tf:              Mitigation; PTE Inversion
Vulnerability Mds:               Vulnerable: Clear CPU buffers attempted, no m
                                 icrocode; SMT Host state unknown
Vulnerability Meltdown:          Mitigation; PTI
Vulnerability Spec store bypass: Vulnerable
Vulnerability Spectre v1:        Mitigation; usercopy/swapgs barriers and __us
                                 er pointer sanitization
Vulnerability Spectre v2:        Mitigation; Full generic retpoline, STIBP dis
                                 abled, RSB filling
Vulnerability Srbds:             Not affected
Vulnerability Tsx async abort:   Not affected
Flags:                           fpu vme de pse tsc msr pae mce cx8 apic sep m
                                 trr pge mca cmov pat pse36 clflush mmx fxsr s
                                 se sse2 ss ht syscall nx pdpe1gb rdtscp lm co
                                 nstant_tsc rep_good nopl xtopology nonstop_ts
                                 c cpuid aperfmperf tsc_known_freq pni pclmulq
                                 dq ssse3 fma cx16 pcid sse4_1 sse4_2 x2apic m
                                 ovbe popcnt tsc_deadline_timer aes xsave avx 
                                 f16c rdrand hypervisor lahf_lm abm 3dnowprefe
                                 tch invpcid_single pti fsgsbase tsc_adjust bm
                                 i1 avx2 smep bmi2 erms invpcid mpx avx512f av
                                 x512dq rdseed adx smap clflushopt clwb avx512
                                 cd avx512bw avx512vl xsaveopt xsavec xgetbv1 
                                 xsaves ida arat pku ospke
```

</details>

<details>
<summary>OS</summary>

```
Linux ip-172-31-28-146 5.4.0-1025-aws
gcc version 7.5.0 (Ubuntu 7.5.0-6ubuntu2)
```
</details>

#### Benchmark
<details>
 <summary>AVX512F - Contiguous only</summary>

metric: gmean, units: ms
|name of test|before_contig_avx512f|after_contig_avx512f|after_contig_avx512f vs before_contig_avx512f|
|---|:-:|:-:|:-:|
|cos::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0011|0.0011|1.01|
|cos::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0018|0.0018|1.01|
|cos::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0033|0.0029|**`1.13`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0011|0.0011|1.01|
|sin::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0018|0.0018|0.98|
|sin::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0032|0.0030|**`1.07`**|

</details>

<details>
 <summary>AVX512F</summary>

metric: gmean, units: ms
|name of test|before_avx512f|after_avx512f|after_avx512f vs before_avx512f|
|---|:-:|:-:|:-:|
|cos::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0011|0.0011|1.01|
|cos::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0018|0.0018|0.99|
|cos::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0034|0.0032|1.05|
|cos::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::2|0.0139|0.0010|**`14.02`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::2|0.0278|0.0019|**`14.76`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::2|0.0561|0.0046|**`12.17`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::3|0.0140|0.0010|**`13.88`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::3|0.0280|0.0019|**`14.76`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::3|0.0565|0.0045|**`12.54`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::10|0.0140|0.0012|**`12.03`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::10|0.0280|0.0020|**`14.18`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::10|0.0562|0.0046|**`12.13`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::1|0.0010|0.0009|**`1.07`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::1|0.0019|0.0017|**`1.08`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::1|0.0038|0.0035|**`1.09`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::2|0.0200|0.0012|**`16.48`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::2|0.0400|0.0025|**`16.22`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::2|0.0799|0.0048|**`16.82`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::3|0.0200|0.0012|**`16.53`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::3|0.0400|0.0024|**`16.88`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::3|0.0800|0.0047|**`17.02`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::10|0.0200|0.0013|**`15.8`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::10|0.0400|0.0025|**`16.08`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::10|0.0801|0.0050|**`16.02`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::1|0.0010|0.0009|**`1.07`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::1|0.0019|0.0017|**`1.09`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::1|0.0039|0.0035|**`1.11`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::2|0.0200|0.0012|**`16.6`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::2|0.0400|0.0024|**`16.65`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::2|0.0802|0.0048|**`16.53`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::3|0.0200|0.0012|**`16.69`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::3|0.0400|0.0024|**`16.92`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::3|0.0802|0.0047|**`17.09`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::10|0.0200|0.0013|**`15.8`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::10|0.0400|0.0025|**`16.01`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::10|0.0804|0.0050|**`16.23`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::1|0.0011|0.0010|**`1.08`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::1|0.0021|0.0019|**`1.11`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::1|0.0042|0.0038|**`1.11`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::2|0.0200|0.0013|**`15.14`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::2|0.0400|0.0026|**`15.54`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::2|0.0801|0.0052|**`15.46`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::3|0.0200|0.0013|**`14.92`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::3|0.0400|0.0026|**`15.54`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::3|0.0799|0.0051|**`15.59`**|
|cos::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::10|0.0200|0.0014|**`14.02`**|
|cos::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::10|0.0400|0.0028|**`14.4`**|
|cos::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::10|0.0802|0.0055|**`14.49`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0011|0.0011|1.01|
|sin::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0017|0.0017|1.03|
|sin::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0033|0.0032|1.02|
|sin::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::2|0.0132|0.0013|**`10.26`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::2|0.0264|0.0020|**`13.36`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::2|0.0533|0.0046|**`11.55`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::3|0.0132|0.0013|**`10.5`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::3|0.0267|0.0020|**`13.49`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::3|0.0532|0.0046|**`11.61`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::10|0.0132|0.0014|**`9.35`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::10|0.0264|0.0021|**`12.63`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::10|0.0528|0.0047|**`11.31`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::1|0.0012|0.0011|1.04|
|sin::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::1|0.0020|0.0019|**`1.06`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::1|0.0038|0.0035|**`1.07`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::2|0.0181|0.0015|**`12.21`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::2|0.0361|0.0023|**`15.41`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::2|0.0723|0.0047|**`15.53`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::3|0.0181|0.0014|**`12.73`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::3|0.0364|0.0023|**`15.76`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::3|0.0723|0.0047|**`15.26`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::10|0.0181|0.0015|**`12.2`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::10|0.0362|0.0024|**`14.85`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::2&nbsp; ->&nbsp; f::10|0.0724|0.0049|**`14.82`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::1|0.0012|0.0011|1.04|
|sin::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::1|0.0020|0.0019|1.05|
|sin::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::1|0.0038|0.0036|**`1.08`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::2|0.0181|0.0015|**`12.45`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::2|0.0362|0.0024|**`15.39`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::2|0.0724|0.0047|**`15.44`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::3|0.0181|0.0014|**`12.65`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::3|0.0365|0.0023|**`15.71`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::3|0.0724|0.0047|**`15.26`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::10|0.0181|0.0015|**`12.29`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::10|0.0362|0.0025|**`14.77`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::3&nbsp; ->&nbsp; f::10|0.0726|0.0049|**`14.92`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::1|0.0013|0.0012|1.04|
|sin::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::1|0.0022|0.0021|**`1.08`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::1|0.0042|0.0038|**`1.09`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::2|0.0181|0.0015|**`11.79`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::2|0.0361|0.0025|**`14.26`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::2|0.0725|0.0051|**`14.28`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::3|0.0181|0.0015|**`11.81`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::3|0.0364|0.0025|**`14.37`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::3|0.0723|0.0052|**`13.91`**|
|sin::1024&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::10|0.0181|0.0016|**`11.17`**|
|sin::2048&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::10|0.0362|0.0027|**`13.24`**|
|sin::4096&nbsp; &nbsp; &nbsp; f::10&nbsp; ->&nbsp; f::10|0.0725|0.0055|**`13.3`**|

</details>

<details>
 <summary>AVX2 & FMA3 - Contiguous only</summary>

metric: gmean, units: ms
|name of test|before_contig_avx2_fma3|after_contig_avx2_fma3|after_contig_avx2_fma3 vs before_contig_avx2_fma3|
|---|:-:|:-:|:-:|
|cos::1024&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0015|0.0014|1.05|
|cos::2048&nbsp; &nbsp; &nbsp; f::1&nbsp; ->&nbsp; f::1|0.0027|0.0026|1.05|
…