@seiko2plus 2020-11-17
the new NPYV intrinsics have moved to separate pull-requests #17790, #17789