@mattip 2020-12-26
Nice speedups. Is this for 32-bit float only or also for 64-bit?