@seiko2plus 2020-12-26
@mattip, just replaced the raw SIMD code of f32 with NPYV.